solve char-short conversion issue after nighttime epiphany
The problem was that sometimes i would successfully combine two chars into a short,
and sometimes the top byte of the short would be 0xff.

The root issue was that during the char-short conversion, if the char were greater
than 0x80, it would be sign-extended during the conversion, causing the new value
to be 0xff80 instead of expected 0x0080.

Solution was to & this new value with 0x00ff, forcing the top byte to 0x00.