I realized i should be using unsigned char/short
this solves the issue i've been having with the top byte of my char to
short conversions becoming 0xff. it was due to c sign extension...
originally i fixed this by doing an & with 0x00ff, but i realized
a better solution is to just use unsigned char/short in the first
place.

i used this to get an accepted answer on stackoverflow lol
https://stackoverflow.com/a/51695560/5924962