Implement pause functionality and general keypress handling
I've created an enumeration in state.h, called COMMAND, which
corresponds to the various commands that can be peformed with a
keypress. I've also documented this in the readme.

Command handling is in the same area of code as general keypress
handling; if a keypress occurs and it is not a chip-8 key, a
function parseCommandKey() is called which returns the corresponding
COMMAND pressed. The code returns from this function to a switch
statement which then will perform an action depending on what command
is called.

The pause functionality feels a bit hacky (and is glitchy) but does
function. It toggles a boolean "paused" and then waits for a new
keypress; when a new keypress occurs the code does a goto to the
top of the switch statement, allowing the emulation to remain paused
as commands are handled.